requirements.txt
peewee==3.16.2
yfinance==0.2.36
boto3==1.34.11
requests==2.32.3
fredapi==0.5.1
python-dateutil==2.8.2
pyspark==3.5.0
delta-spark==3.0.0
aws-cdk-lib~=2.115.0
constructs>=10.0.0,<11.0.0